The shotgun approach to object cleanup: Destroy everything!
Up to now, I had problems with "ghost walls". Turns out that these are
the objects from the previous games (i.e. the intro game), which are not
correctly cleaned up by the KolfGame dtor. Because the CanvasItems have
so many interdependencies and stuff, I'll just delete items with a shape
for now. This needs to be replaced by proper cleanup somewhen.

svn path=/trunk/KDE/kdegames/kolf/; revision=1191722